docker-build/build-script-source/requirements.txt
colorama==0.4.1
PyYAML==5.1.1
